+++ tests/CommunityToolkit.Mvvm.UnitTests/Collections/Test_ObservableGroup.cs
@@ -3,43 +3,39 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.ObjectModel;
-using System.Collections.Specialized;
 using System.Linq;
 using CommunityToolkit.Mvvm.Collections;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace CommunityToolkit.Mvvm.UnitTests;
+namespace CommunityToolkit.Mvvm.UnitTests.Collections;
 
 [TestClass]
-public class ReadOnlyObservableGroupTests
+public class Test_ObservableGroup
 {
     [TestMethod]
-    public void Ctor_WithKeyAndOBservableCollection_ShouldHaveExpectedInitialState()
+    public void Ctor_ShouldHaveExpectedState()
     {
-        ObservableCollection<int> source = new(new[] { 1, 2, 3 });
-        ReadOnlyObservableGroup<string, int> group = new("key", source);
+        ObservableGroup<string, int> group = new("key");
+
+        Assert.AreEqual(group.Key, "key");
+        Assert.AreEqual(group.Count, 0);
+    }
+
+    [TestMethod]
+    public void Ctor_WithGrouping_ShouldHaveExpectedState()
+    {
+        IntGroup source = new("key", new[] { 1, 2, 3 });
+        ObservableGroup<string, int> group = new(source);
 
         Assert.AreEqual(group.Key, "key");
         CollectionAssert.AreEqual(group, new[] { 1, 2, 3 });
     }
 
     [TestMethod]
-    public void Ctor_ObservableGroup_ShouldHaveExpectedInitialState()
+    public void Ctor_WithCollection_ShouldHaveExpectedState()
     {
         int[] source = new[] { 1, 2, 3 };
-        ObservableGroup<string, int> sourceGroup = new("key", source);
-        ReadOnlyObservableGroup<string, int> group = new(sourceGroup);
-
-        Assert.AreEqual(group.Key, "key");
-        CollectionAssert.AreEqual(group, new[] { 1, 2, 3 });
-    }
-
-    [TestMethod]
-    public void Ctor_WithKeyAndCollection_ShouldHaveExpectedInitialState()
-    {
-        ObservableCollection<int> source = new() { 1, 2, 3 };
-        ReadOnlyObservableGroup<string, int> group = new("key", source);
+        ObservableGroup<string, int> group = new("key", source);
 
         Assert.AreEqual(group.Key, "key");
         CollectionAssert.AreEqual(group, new[] { 1, 2, 3 });
@@ -50,16 +46,14 @@ public class ReadOnlyObservableGroupTests
     {
         bool collectionChangedEventRaised = false;
         int[] source = new[] { 1, 2, 3 };
-        ObservableGroup<string, int> sourceGroup = new("key", source);
-        ReadOnlyObservableGroup<string, int> group = new(sourceGroup);
+        ObservableGroup<string, int> group = new("key", source);
 
-        ((INotifyCollectionChanged)group).CollectionChanged += (s, e) => collectionChangedEventRaised = true;
+        group.CollectionChanged += (s, e) => collectionChangedEventRaised = true;
 
-        sourceGroup.Add(4);
+        group.Add(4);
 
         Assert.AreEqual(group.Key, "key");
         CollectionAssert.AreEqual(group, new[] { 1, 2, 3, 4 });
-
         Assert.IsTrue(collectionChangedEventRaised);
     }
 
@@ -68,16 +62,14 @@ public class ReadOnlyObservableGroupTests
     {
         bool collectionChangedEventRaised = false;
         int[] source = new[] { 1, 2, 3 };
-        ObservableGroup<string, int> sourceGroup = new("key", source);
-        ReadOnlyObservableGroup<string, int> group = new(sourceGroup);
+        ObservableGroup<string, int> group = new("key", source);
 
-        ((INotifyCollectionChanged)group).CollectionChanged += (s, e) => collectionChangedEventRaised = true;
+        group.CollectionChanged += (s, e) => collectionChangedEventRaised = true;
 
-        sourceGroup[1] = 4;
+        group[1] = 4;
 
         Assert.AreEqual(group.Key, "key");
         CollectionAssert.AreEqual(group, new[] { 1, 4, 3 });
-
         Assert.IsTrue(collectionChangedEventRaised);
     }
 
@@ -86,16 +78,14 @@ public class ReadOnlyObservableGroupTests
     {
         bool collectionChangedEventRaised = false;
         int[] source = new[] { 1, 2, 3 };
-        ObservableGroup<string, int> sourceGroup = new("key", source);
-        ReadOnlyObservableGroup<string, int> group = new(sourceGroup);
+        ObservableGroup<string, int>? group = new("key", source);
 
-        ((INotifyCollectionChanged)group).CollectionChanged += (s, e) => collectionChangedEventRaised = true;
+        group.CollectionChanged += (s, e) => collectionChangedEventRaised = true;
 
-        _ = sourceGroup.Remove(1);
+        _ = group.Remove(1);
 
         Assert.AreEqual(group.Key, "key");
         CollectionAssert.AreEqual(group, new[] { 2, 3 });
-
         Assert.IsTrue(collectionChangedEventRaised);
     }
 
@@ -104,16 +94,14 @@ public class ReadOnlyObservableGroupTests
     {
         bool collectionChangedEventRaised = false;
         int[] source = new[] { 1, 2, 3 };
-        ObservableGroup<string, int> sourceGroup = new("key", source);
-        ReadOnlyObservableGroup<string, int> group = new(sourceGroup);
+        ObservableGroup<string, int>? group = new("key", source);
 
-        ((INotifyCollectionChanged)group).CollectionChanged += (s, e) => collectionChangedEventRaised = true;
+        group.CollectionChanged += (s, e) => collectionChangedEventRaised = true;
 
-        sourceGroup.Clear();
+        group.Clear();
 
         Assert.AreEqual(group.Key, "key");
-        CollectionAssert.AreEqual(group, Array.Empty<int>());
-
+        Assert.AreEqual(group.Count, 0);
         Assert.IsTrue(collectionChangedEventRaised);
     }
 
@@ -122,8 +110,7 @@ public class ReadOnlyObservableGroupTests
     [DataRow(3)]
     public void IReadOnlyObservableGroup_ShouldReturnExpectedValues(int count)
     {
-        ObservableGroup<string, int> sourceGroup = new("key", Enumerable.Range(0, count));
-        ReadOnlyObservableGroup<string, int> group = new(sourceGroup);
+        ObservableGroup<string, int> group = new("key", Enumerable.Range(0, count));
         IReadOnlyObservableGroup iReadOnlyObservableGroup = group;
 
         Assert.AreEqual(iReadOnlyObservableGroup.Key, "key");
@@ -132,22 +119,38 @@ public class ReadOnlyObservableGroupTests
 
     [TestMethod]
     [ExpectedException(typeof(ArgumentNullException))]
-    public void Ctor_NullKeyWithNotNullElements()
+    public void Ctor_NullKey()
     {
-        _ = new ReadOnlyObservableGroup<string, int>(null!, new ObservableCollection<int>());
-    }
-
-    [TestMethod]
-    [ExpectedException(typeof(ArgumentNullException))]
-    public void Ctor_NotNullKeyWithNullElements()
-    {
-        _ = new ReadOnlyObservableGroup<string, int>("A", null!);
+        _ = new ObservableGroup<string, int>((string)null!);
     }
 
     [TestMethod]
     [ExpectedException(typeof(ArgumentNullException))]
     public void Ctor_NullGroup()
     {
-        _ = new ReadOnlyObservableGroup<string, int>(null!);
+        _ = new ObservableGroup<string, int>((IGrouping<string, int>)null!);
+    }
+
+    [TestMethod]
+    [ExpectedException(typeof(ArgumentNullException))]
+    public void Ctor_NullKeyWithNotNullElements()
+    {
+        _ = new ObservableGroup<string, int>(null!, new int[0]);
+    }
+
+    [TestMethod]
+    [ExpectedException(typeof(ArgumentNullException))]
+    public void Ctor_NotNullKeyWithNullElements()
+    {
+        _ = new ObservableGroup<string, int>("A", null!);
+    }
+
+    [TestMethod]
+    [ExpectedException(typeof(ArgumentNullException))]
+    public void Ctor_NullKeySetter()
+    {
+        ObservableGroup<string, int> group = new("A");
+
+        group.Key = null!;
     }
 }
